renaming autoregressive_decoder to autoregressive
