+++ neuralmonkey/decoders/autoregressive.py
